move the stack to vif/netstack
